Added stats graph and favorite

+++ app/src/main/java/com/example/pokemonapp/api/response/models/StatsModel.kt
@@ -3,5 +3,5 @@ package com.example.pokemonapp.api.response.models
 import com.google.gson.annotations.SerializedName
 
 data class StatsModel(
-    @SerializedName("base_stat") val baseStat: Int
+    @SerializedName("base_stat") val baseStat: Float
 )
